build: 修改配置文件
