+++ .eslintrc.js
@@ -26,6 +26,7 @@ module.exports = {
   rules: {
     '@typescript-eslint/no-explicit-any': ['off'],
     'react/react-in-jsx-scope': 'off',
+    '@typescript-eslint/no-var-requires': 'off'
     // "react/jsx-uses-react": "off",
   },
   ignorePatterns: [
