+++ demo/index.tsx
@@ -1,7 +1,7 @@
 import { createRoot } from 'react-dom/client';
 import { useState, useEffect } from 'react';
 import { getDataList } from './data';
-import VirtualScrollList from '../src/index';
+import VirtualScrollList from '../src/virtual-scroll-list/index';
 // import VirtualScrollList from '../dist/index';
 // import VirtualScrollList from "virtual-scroll-list_ferying";
 import './index.scss';
