--- src/index-1.tsx
@@ -1,2 +1,0 @@
-
-// export { default as VirtualScrollList} from './virtual-scroll-list'
